# app/Http/Controllers/ContingentController.php
<?php

namespace App\Http\Controllers;

use OpenApi\Attributes as OA;

use App\Models\Contingent;
use App\Models\Player;
use App\Models\User;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\DB;

class ContingentController extends Controller
{
    #[OA\Get(
        path: "/api/contingents",
        operationId: "listContingents",
        tags: ["Contingents"],
        summary: "Daftar semua kontingen",
        description: "Mengembalikan semua kontingen beserta informasi PIC dan jumlah anggota. Dapat diakses publik."
    )]
    #[OA\Response(
        response: 200,
        description: "Berhasil",
        content: new OA\JsonContent(
            properties: [
                new OA\Property(property: "status", type: "string", example: "success"),
                new OA\Property(property: "data", type: "array", items: new OA\Items(
                    properties: [
                        new OA\Property(property: "id", type: "integer", example: 1),
                        new OA\Property(property: "name", type: "string", example: "Fakultas Informatika"),
                        new OA\Property(property: "pic", type: "object", nullable: true),
                        new OA\Property(property: "players_count", type: "integer", example: 12),
                    ]
                )),
            ]
        )
    )]
    public function index()
    {
        $contingents = Contingent::with(['pic:id,name,email'])
            ->withCount('players')
            ->get();

        return response()->json(['status' => 'success', 'data' => $contingents]);
    }

    #[OA\Get(
        path: "/api/contingents/{id}",
        operationId: "showContingent",
        tags: ["Contingents"],
        summary: "Detail kontingen beserta anggota",
        description: "Mengembalikan detail kontingen termasuk daftar seluruh player yang terdaftar."
    )]
    #[OA\Parameter(name: "id", in: "path", required: true, schema: new OA\Schema(type: "integer"))]
    #[OA\Response(
        response: 200,
        description: "Berhasil",
        content: new OA\JsonContent(
            properties: [
                new OA\Property(property: "status", type: "string", example: "success"),
                new OA\Property(property: "data",   ref: "#/components/schemas/ContingentObject"),
            ]
        )
    )]
    #[OA\Response(response: 404, description: "Kontingen tidak ditemukan", content: new OA\JsonContent(ref: "#/components/schemas/ErrorResponse"))]
    public function show($id)
    {
        $contingent = Contingent::with([
            'pic:id,name,email',
            'players:id,contingent_id,name,nim_nip,sport_id,sport_category_id,photo_path,verification_status',
        ])->findOrFail($id);

        return response()->json(['status' => 'success', 'data' => $contingent]);
    }

    #[OA\Post(
        path: "/api/contingents",
        operationId: "createContingent",
        tags: ["Contingents"],
        summary: "Buat kontingen baru",
        description: "Admin/panitia membuat kontingen baru yang mewakili satu fakultas atau divisi.",
        security: [["bearerAuth" => []]]
    )]
    #[OA\RequestBody(
        required: true,
        content: new OA\JsonContent(
            required: ["name"],
            properties: [
                new OA\Property(property: "name", type: "string", example: "Fakultas Informatika"),
            ]
        )
    )]
    #[OA\Response(
        response: 201,
        description: "Kontingen berhasil dibuat",
        content: new OA\JsonContent(
            properties: [
                new OA\Property(property: "status",  type: "string", example: "success"),
                new OA\Property(property: "message", type: "string", example: "Kontingen berhasil dibuat."),
                new OA\Property(property: "data",    ref: "#/components/schemas/ContingentObject"),
            ]
        )
    )]
    #[OA\Response(response: 422, description: "Nama sudah digunakan", content: new OA\JsonContent(ref: "#/components/schemas/ErrorResponse"))]
    public function store(Request $request)
    {
        $validated = $request->validate([
            'name' => 'required|string|max:255|unique:contingents,name',
        ]);

        $contingent = Contingent::create(['name' => $validated['name']]);

        return response()->json([
            'status'  => 'success',
            'message' => 'Kontingen berhasil dibuat.',
            'data'    => $contingent,
        ], 201);
    }

    #[OA\Put(
        path: "/api/contingents/{id}",
        operationId: "updateContingent",
        tags: ["Contingents"],
        summary: "Update nama kontingen",
        description: "Admin/panitia mengupdate nama kontingen.",
        security: [["bearerAuth" => []]]
    )]
    #[OA\Parameter(name: "id", in: "path", required: true, schema: new OA\Schema(type: "integer"))]
    #[OA\RequestBody(
        required: true,
        content: new OA\JsonContent(
            required: ["name"],
            properties: [new OA\Property(property: "name", type: "string", example: "Fakultas Teknik")]
        )
    )]
    #[OA\Response(
        response: 200,
        description: "Berhasil diperbarui",
        content: new OA\JsonContent(
            properties: [
                new OA\Property(property: "status",  type: "string", example: "success"),
                new OA\Property(property: "message", type: "string", example: "Kontingen berhasil diperbarui."),
                new OA\Property(property: "data",    ref: "#/components/schemas/ContingentObject"),
            ]
        )
    )]
    #[OA\Response(response: 404, description: "Kontingen tidak ditemukan", content: new OA\JsonContent(ref: "#/components/schemas/ErrorResponse"))]
    public function update(Request $request, $id)
    {
        $contingent = Contingent::findOrFail($id);

        $validated = $request->validate([
            'name' => 'required|string|max:255|unique:contingents,name,' . $contingent->id,
        ]);

        $contingent->update(['name' => $validated['name']]);

        return response()->json([
            'status'  => 'success',
            'message' => 'Kontingen berhasil diperbarui.',
            'data'    => $contingent->fresh()->load('pic:id,name,email'),
        ]);
    }

    #[OA\Delete(
        path: "/api/contingents/{id}",
        operationId: "deleteContingent",
        tags: ["Contingents"],
        summary: "Hapus kontingen",
        description: "Admin/panitia menghapus kontingen. Player yang terhubung akan memiliki contingent_id = null.",
        security: [["bearerAuth" => []]]
    )]
    #[OA\Parameter(name: "id", in: "path", required: true, schema: new OA\Schema(type: "integer"))]
    #[OA\Response(response: 200, description: "Berhasil dihapus", content: new OA\JsonContent(ref: "#/components/schemas/SuccessMessage"))]
    #[OA\Response(response: 404, description: "Kontingen tidak ditemukan", content: new OA\JsonContent(ref: "#/components/schemas/ErrorResponse"))]
    public function destroy($id)
    {
        $contingent = Contingent::findOrFail($id);
        $contingent->delete();

        return response()->json(['status' => 'success', 'message' => 'Kontingen berhasil dihapus.']);
    }

    #[OA\Put(
        path: "/api/contingents/{id}/assign-pic",
        operationId: "assignPic",
        tags: ["Contingents"],
        summary: "Tugaskan PIC kontingen",
        description: "Admin/panitia menugaskan seorang user sebagai PIC (Person In Charge) dari kontingen. Role user otomatis diubah menjadi `pic_kontingen` dan player-nya dimasukkan ke kontingen tersebut.",
        security: [["bearerAuth" => []]]
    )]
    #[OA\Parameter(name: "id", in: "path", required: true, schema: new OA\Schema(type: "integer"))]
    #[OA\RequestBody(
        required: true,
        content: new OA\JsonContent(
            required: ["user_id"],
            properties: [
                new OA\Property(property: "user_id", type: "integer", example: 5, description: "ID user yang akan dijadikan PIC"),
            ]
        )
    )]
    #[OA\Response(
        response: 200,
        description: "PIC berhasil ditugaskan",
        content: new OA\JsonContent(
            properties: [
                new OA\Property(property: "status",  type: "string", example: "success"),
                new OA\Property(property: "message", type: "string", example: "Budi berhasil ditugaskan sebagai PIC kontingen Fakultas Informatika."),
                new OA\Property(property: "data",    ref: "#/components/schemas/ContingentObject"),
            ]
        )
    )]
    #[OA\Response(response: 422, description: "User adalah panitia/admin", content: new OA\JsonContent(ref: "#/components/schemas/ErrorResponse"))]
    public function assignPic(Request $request, $id)
    {
        $contingent = Contingent::findOrFail($id);

        $validated = $request->validate([
            'user_id' => 'required|integer|exists:users,id',
        ]);

        $user = User::findOrFail($validated['user_id']);

        if ($user->role === 'panitia' || $user->role === 'admin') {
            return response()->json([
                'status'  => 'error',
                'message' => 'Tidak dapat menugaskan panitia atau admin sebagai PIC kontingen.',
            ], 422);
        }

        DB::beginTransaction();
        try {
            $user->update(['role' => 'pic_kontingen']);
            $contingent->update(['pic_user_id' => $user->id]);

            if ($user->player) {
                $user->player->update(['contingent_id' => $contingent->id]);
            }

            DB::commit();

            return response()->json([
                'status'  => 'success',
                'message' => "{$user->name} berhasil ditugaskan sebagai PIC kontingen {$contingent->name}.",
                'data'    => $contingent->fresh()->load('pic:id,name,email'),
            ]);
        } catch (\Throwable $e) {
            DB::rollBack();
            return response()->json(['status' => 'error', 'message' => $e->getMessage()], 500);
        }
    }

    #[OA\Get(
        path: "/api/contingents/my",
        operationId: "myContingent",
        tags: ["Contingents"],
        summary: "Detail kontingen milik PIC yang login",
        description: "PIC kontingen melihat detail kontingennya sendiri beserta seluruh anggota.",
        security: [["bearerAuth" => []]]
    )]
    #[OA\Response(
        response: 200,
        description: "Berhasil",
        content: new OA\JsonContent(
            properties: [
                new OA\Property(property: "status", type: "string", example: "success"),
                new OA\Property(property: "data",   ref: "#/components/schemas/ContingentObject"),
            ]
        )
    )]
    #[OA\Response(response: 404, description: "Belum ditugaskan sebagai PIC", content: new OA\JsonContent(ref: "#/components/schemas/ErrorResponse"))]
    public function myContingent(Request $request)
    {
        $contingent = $request->user()->managedContingent;

        if (!$contingent) {
            return response()->json([
                'status'  => 'error',
                'message' => 'Anda belum ditugaskan sebagai PIC kontingen manapun.',
            ], 404);
        }

        $contingent->load([
            'players:id,contingent_id,name,nim_nip,sport_id,sport_category_id,photo_path,verification_status',
        ]);

        return response()->json(['status' => 'success', 'data' => $contingent]);
    }

    #[OA\Post(
        path: "/api/contingents/my/players",
        operationId: "addPlayerToContingent",
        tags: ["Contingents"],
        summary: "PIC menambahkan player ke kontingen",
        description: "PIC kontingen menambahkan seorang player ke dalam kontingennya. Player harus sudah memiliki akun di sistem.",
        security: [["bearerAuth" => []]]
    )]
    #[OA\RequestBody(
        required: true,
        content: new OA\JsonContent(
            required: ["player_id"],
            properties: [
                new OA\Property(property: "player_id", type: "integer", example: 7),
            ]
        )
    )]
    #[OA\Response(
        response: 200,
        description: "Player berhasil ditambahkan",
        content: new OA\JsonContent(
            properties: [
                new OA\Property(property: "status",  type: "string", example: "success"),
                new OA\Property(property: "message", type: "string", example: "Ahmad Fauzi berhasil ditambahkan ke kontingen Fakultas Informatika."),
                new OA\Property(property: "data",    ref: "#/components/schemas/PlayerObject"),
            ]
        )
    )]
    #[OA\Response(response: 422, description: "Player sudah ada di kontingen lain", content: new OA\JsonContent(ref: "#/components/schemas/ErrorResponse"))]
    #[OA\Response(response: 403, description: "Belum ditugaskan sebagai PIC", content: new OA\JsonContent(ref: "#/components/schemas/ErrorResponse"))]
    public function addPlayer(Request $request)
    {
        $user       = $request->user();
        $contingent = $user->managedContingent;

        if (!$contingent) {
            return response()->json(['status' => 'error', 'message' => 'Anda belum ditugaskan sebagai PIC kontingen manapun.'], 403);
        }

        $validated = $request->validate([
            'player_id' => 'required|integer|exists:players,id',
        ]);

        $player = Player::findOrFail($validated['player_id']);

        if ($player->contingent_id && $player->contingent_id !== $contingent->id) {
            return response()->json(['status' => 'error', 'message' => 'Player ini sudah terdaftar di kontingen lain.'], 422);
        }

        $player->update(['contingent_id' => $contingent->id]);

        return response()->json([
            'status'  => 'success',
            'message' => "{$player->name} berhasil ditambahkan ke kontingen {$contingent->name}.",
            'data'    => $player->fresh(['contingent']),
        ]);
    }

    #[OA\Delete(
        path: "/api/contingents/my/players/{player_id}",
        operationId: "removePlayerFromContingent",
        tags: ["Contingents"],
        summary: "PIC mengeluarkan player dari kontingen",
        description: "PIC kontingen mengeluarkan seorang player dari kontingennya. PIC tidak dapat mengeluarkan dirinya sendiri.",
        security: [["bearerAuth" => []]]
    )]
    #[OA\Parameter(name: "player_id", in: "path", required: true, schema: new OA\Schema(type: "integer"))]
    #[OA\Response(response: 200, description: "Player berhasil dikeluarkan", content: new OA\JsonContent(ref: "#/components/schemas/SuccessMessage"))]
    #[OA\Response(response: 403, description: "Player bukan anggota kontingen ini", content: new OA\JsonContent(ref: "#/components/schemas/ErrorResponse"))]
    public function removePlayer(Request $request, $playerId)
    {
        $user       = $request->user();
        $contingent = $user->managedContingent;

        if (!$contingent) {
            return response()->json(['status' => 'error', 'message' => 'Anda belum ditugaskan sebagai PIC kontingen manapun.'], 403);
        }

        $player = Player::findOrFail($playerId);

        if ($player->contingent_id !== $contingent->id) {
            return response()->json(['status' => 'error', 'message' => 'Player ini bukan anggota kontingen Anda.'], 403);
        }

        if ($player->user_id === $user->id) {
            return response()->json(['status' => 'error', 'message' => 'PIC tidak dapat dikeluarkan dari kontingen melalui endpoint ini.'], 422);
        }

        $player->update(['contingent_id' => null]);

        return response()->json([
            'status'  => 'success',
            'message' => "{$player->name} berhasil dikeluarkan dari kontingen {$contingent->name}.",
        ]);
    }
}