# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'TelU Cup Backend API'
  description: 'Dokumentasi lengkap API manajemen kompetisi internal Tel-U Cup. Mencakup autentikasi, manajemen cabang olahraga, kontingen, registrasi tim, dan bagan pertandingan gugur.'
  version: 2.0.0
servers:
  -
    url: 'http://localhost:8000'
    description: 'Development Server'
paths:
...
  /api/contingents:
    get:
      tags:
        - Contingents
      summary: 'Daftar semua kontingen'
      description: 'Mengembalikan semua kontingen beserta informasi PIC dan jumlah anggota. Dapat diakses publik.'
      operationId: listContingents
      responses:
        200:
          description: Berhasil
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      properties:
                        id:
                          type: integer
                          example: 1
                        name:
                          type: string
                          example: 'Fakultas Informatika'
                        pic:
                          type: object
                          nullable: true
                        players_count:
                          type: integer
                          example: 12
                      type: object
                type: object
    post:
      tags:
        - Contingents
      summary: 'Buat kontingen baru'
      description: 'Admin/panitia membuat kontingen baru yang mewakili satu fakultas atau divisi.'
      operationId: createContingent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
              properties:
                name:
                  type: string
                  example: 'Fakultas Informatika'
              type: object
      responses:
        201:
          description: 'Kontingen berhasil dibuat'
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: 'Kontingen berhasil dibuat.'
                  data:
                    $ref: '#/components/schemas/ContingentObject'
                type: object
        422:
          description: 'Nama sudah digunakan'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          bearerAuth: []
  '/api/contingents/{id}':
    get:
      tags:
        - Contingents
      summary: 'Detail kontingen beserta anggota'
      description: 'Mengembalikan detail kontingen termasuk daftar seluruh player yang terdaftar.'
      operationId: showContingent
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Berhasil
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/ContingentObject'
                type: object
        404:
          description: 'Kontingen tidak ditemukan'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Contingents
      summary: 'Update nama kontingen'
      description: 'Admin/panitia mengupdate nama kontingen.'
      operationId: updateContingent
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
              properties:
                name:
                  type: string
                  example: 'Fakultas Teknik'
              type: object
      responses:
        200:
          description: 'Berhasil diperbarui'
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: 'Kontingen berhasil diperbarui.'
                  data:
                    $ref: '#/components/schemas/ContingentObject'
                type: object
        404:
          description: 'Kontingen tidak ditemukan'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Contingents
      summary: 'Hapus kontingen'
      description: 'Admin/panitia menghapus kontingen. Player yang terhubung akan memiliki contingent_id = null.'
      operationId: deleteContingent
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Berhasil dihapus'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        404:
          description: 'Kontingen tidak ditemukan'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          bearerAuth: []
  '/api/contingents/{id}/assign-pic':
    put:
      tags:
        - Contingents
      summary: 'Tugaskan PIC kontingen'
      description: 'Admin/panitia menugaskan seorang user sebagai PIC (Person In Charge) dari kontingen. Role user otomatis diubah menjadi `pic_kontingen` dan player-nya dimasukkan ke kontingen tersebut.'
      operationId: assignPic
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - user_id
              properties:
                user_id:
                  description: 'ID user yang akan dijadikan PIC'
                  type: integer
                  example: 5
              type: object
      responses:
        200:
          description: 'PIC berhasil ditugaskan'
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: 'Budi berhasil ditugaskan sebagai PIC kontingen Fakultas Informatika.'
                  data:
                    $ref: '#/components/schemas/ContingentObject'
                type: object
        422:
          description: 'User adalah panitia/admin'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          bearerAuth: []
  /api/contingents/my:
    get:
      tags:
        - Contingents
      summary: 'Detail kontingen milik PIC yang login'
      description: 'PIC kontingen melihat detail kontingennya sendiri beserta seluruh anggota.'
      operationId: myContingent
      responses:
        200:
          description: Berhasil
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/ContingentObject'
                type: object
        404:
          description: 'Belum ditugaskan sebagai PIC'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          bearerAuth: []
  /api/contingents/my/players:
    post:
      tags:
        - Contingents
      summary: 'PIC menambahkan player ke kontingen'
      description: 'PIC kontingen menambahkan seorang player ke dalam kontingennya. Player harus sudah memiliki akun di sistem.'
      operationId: addPlayerToContingent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - player_id
              properties:
                player_id:
                  type: integer
                  example: 7
              type: object
      responses:
        200:
          description: 'Player berhasil ditambahkan'
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: 'Ahmad Fauzi berhasil ditambahkan ke kontingen Fakultas Informatika.'
                  data:
                    $ref: '#/components/schemas/PlayerObject'
                type: object
        422:
          description: 'Player sudah ada di kontingen lain'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        403:
          description: 'Belum ditugaskan sebagai PIC'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          bearerAuth: []
  '/api/contingents/my/players/{player_id}':
    delete:
      tags:
        - Contingents
      summary: 'PIC mengeluarkan player dari kontingen'
      description: 'PIC kontingen mengeluarkan seorang player dari kontingennya. PIC tidak dapat mengeluarkan dirinya sendiri.'
      operationId: removePlayerFromContingent
      parameters:
        -
          name: player_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Player berhasil dikeluarkan'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        403:
          description: 'Player bukan anggota kontingen ini'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          bearerAuth: []
...
components:
  schemas:
    ErrorResponse:
      description: 'Format respon error standar'
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          example: 'Pesan error yang menjelaskan penyebab kegagalan.'
      type: object
    SuccessMessage:
      description: 'Format respon sukses tanpa data'
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: 'Operasi berhasil.'
      type: object
    UserObject:
      description: 'Data akun pengguna'
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'Budi Santoso'
        email:
          type: string
          format: email
          example: budi@telkomuniversity.ac.id
        role:
          type: string
          example: player
          enum:
            - admin
            - panitia
            - pic_kontingen
            - player
        is_kacamata:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
...
    ContingentObject:
      description: 'Kontingen yang mewakili satu fakultas atau divisi'
      properties:
        id:
          type: integer
          example: 3
        name:
          type: string
          example: 'Fakultas Informatika'
        pic_user_id:
          type: integer
          example: 5
          nullable: true
        pic:
          oneOf:
            -
              $ref: '#/components/schemas/UserObject'
          nullable: true
        players_count:
          description: 'Hanya tersedia pada endpoint list'
          type: integer
          example: 12
        created_at:
          type: string
          format: date-time
      type: object
    PlayerObject:
      description: 'Data pemain (athlete)'
      properties:
        id:
          type: integer
          example: 7
        user_id:
          type: integer
          example: 10
        name:
          type: string
          example: 'Ahmad Fauzi'
        nim_nip:
          type: string
          example: '1301234567'
          nullable: true
        sport_id:
          type: integer
          example: 1
          nullable: true
        sport_category_id:
          type: integer
          example: 2
          nullable: true
        contingent_id:
          type: integer
          example: 3
          nullable: true
        photo_path:
          type: string
          example: 'https://res.cloudinary.com/...'
          nullable: true
        verification_status:
          type: string
          example: verified
          nullable: true
        checked_in_at:
          type: string
          format: date-time
          nullable: true
        employee_status:
          type: string
          example: Mahasiswa
          nullable: true
        work_location:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
...